+++ test/knapsack_test.jl
@@ -19,8 +19,8 @@ end
     @testset "exact_knapsack" begin
         @testset "classic example" begin
             W = 10
-            weights = Int64[2, 3, 4, 5]
-            values  = Int64[3, 4, 5, 6]
+            weights = Int[2, 3, 4, 5]
+            values  = Int[3, 4, 5, 6]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 13
             validate_knapsack_solution(W, weights, values, chosen, opt)
@@ -28,17 +28,17 @@ end
 
         @testset "single item fits" begin
             W = 5
-            weights = Int64[3]
-            values  = Int64[10]
+            weights = Int[3]
+            values  = Int[10]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 10
-            @test chosen == Int64[1]
+            @test chosen == Int[1]
         end
 
         @testset "single item too heavy" begin
             W = 2
-            weights = Int64[5]
-            values  = Int64[10]
+            weights = Int[5]
+            values  = Int[10]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 0
             @test isempty(chosen)
@@ -46,17 +46,17 @@ end
 
         @testset "all items fit" begin
             W = 100
-            weights = Int64[10, 20, 30]
-            values  = Int64[5, 10, 15]
+            weights = Int[10, 20, 30]
+            values  = Int[5, 10, 15]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 30
-            @test sort(chosen) == Int64[1, 2, 3]
+            @test sort(chosen) == Int[1, 2, 3]
         end
 
         @testset "no items fit" begin
             W = 5
-            weights = Int64[10, 20, 30]
-            values  = Int64[100, 200, 300]
+            weights = Int[10, 20, 30]
+            values  = Int[100, 200, 300]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 0
             @test isempty(chosen)
@@ -64,8 +64,8 @@ end
 
         @testset "empty input" begin
             W = 10
-            weights = Int64[]
-            values  = Int64[]
+            weights = Int[]
+            values  = Int[]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 0
             @test isempty(chosen)
@@ -73,8 +73,8 @@ end
 
         @testset "tie in value different weights" begin
             W = 10
-            weights = Int64[5, 8, 10]
-            values  = Int64[10, 10, 10]
+            weights = Int[5, 8, 10]
+            values  = Int[10, 10, 10]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 10
             validate_knapsack_solution(W, weights, values, chosen, opt)
@@ -82,8 +82,8 @@ end
 
         @testset "optimal subset selection" begin
             W = 50
-            weights = Int64[10, 20, 30]
-            values  = Int64[60, 100, 120]
+            weights = Int[10, 20, 30]
+            values  = Int[60, 100, 120]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 220
             validate_knapsack_solution(W, weights, values, chosen, opt)
@@ -91,8 +91,8 @@ end
 
         @testset "multiple items same efficiency" begin
             W = 15
-            weights = Int64[5, 5, 5, 5]
-            values  = Int64[10, 10, 10, 10]
+            weights = Int[5, 5, 5, 5]
+            values  = Int[10, 10, 10, 10]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 30
             validate_knapsack_solution(W, weights, values, chosen, opt)
@@ -100,8 +100,8 @@ end
 
         @testset "greedy fails but DP succeeds" begin
             W = 10
-            weights = Int64[1, 2, 3, 4, 5]
-            values  = Int64[1, 6, 10, 16, 21]
+            weights = Int[1, 2, 3, 4, 5]
+            values  = Int[1, 6, 10, 16, 21]
             opt, chosen = exact_knapsack(W, weights, values)
             @test opt == 38  
             validate_knapsack_solution(W, weights, values, chosen, opt)
@@ -111,8 +111,8 @@ end
     @testset "ptas_knapsack" begin
         @testset "classic example with tight epsilon" begin
             W = 10
-            weights = Int64[2, 3, 4, 5]
-            values  = Int64[3, 4, 5, 6]
+            weights = Int[2, 3, 4, 5]
+            values  = Int[3, 4, 5, 6]
             epsilon = 0.1
             opt_approx, chosen = ptas_knapsack(W, epsilon, weights, values)
             opt_exact, _ = exact_knapsack(W, weights, values)
@@ -122,18 +122,18 @@ end
 
         @testset "single item" begin
             W = 5
-            weights = Int64[3]
-            values  = Int64[10]
+            weights = Int[3]
+            values  = Int[10]
             epsilon = 0.2
             opt, chosen = ptas_knapsack(W, epsilon, weights, values)
             @test opt == 10
-            @test chosen == Int64[1]
+            @test chosen == Int[1]
         end
 
         @testset "empty input" begin
             W = 10
-            weights = Int64[]
-            values  = Int64[]
+            weights = Int[]
+            values  = Int[]
             epsilon = 0.1
             opt, chosen = ptas_knapsack(W, epsilon, weights, values)
             @test opt == 0
@@ -142,8 +142,8 @@ end
 
         @testset "approximation quality various epsilon" begin
             W = 50
-            weights = Int64[10, 20, 30, 15, 25]
-            values  = Int64[60, 100, 120, 80, 110]
+            weights = Int[10, 20, 30, 15, 25]
+            values  = Int[60, 100, 120, 80, 110]
             opt_exact, _ = exact_knapsack(W, weights, values)
             
             for epsilon in [0.5, 0.3, 0.1, 0.05]
@@ -155,8 +155,8 @@ end
 
         @testset "zero scaling fallback" begin
             W = 10
-            weights = Int64[1, 2, 3]
-            values  = Int64[1, 1, 1]
+            weights = Int[1, 2, 3]
+            values  = Int[1, 1, 1]
             epsilon = 10.0  # huge epsilon forces fallback
             opt, chosen = ptas_knapsack(W, epsilon, weights, values)
             opt_exact, _ = exact_knapsack(W, weights, values)
@@ -166,8 +166,8 @@ end
 
         @testset "negative epsilon fallback" begin
             W = 10
-            weights = Int64[2, 3, 4]
-            values  = Int64[3, 4, 5]
+            weights = Int[2, 3, 4]
+            values  = Int[3, 4, 5]
             epsilon = -0.1
             opt, chosen = ptas_knapsack(W, epsilon, weights, values)
             opt_exact, _ = exact_knapsack(W, weights, values)
@@ -177,8 +177,8 @@ end
 
         @testset "all scaled to zero fallback" begin
             W = 100
-            weights = Int64[10, 20, 30]
-            values  = Int64[1, 2, 3]
+            weights = Int[10, 20, 30]
+            values  = Int[1, 2, 3]
             epsilon = 0.9  # high epsilon, small values
             opt, chosen = ptas_knapsack(W, epsilon, weights, values)
             @test opt >= 0
@@ -187,8 +187,8 @@ end
 
         @testset "identical values" begin
             W = 20
-            weights = Int64[5, 10, 15]
-            values  = Int64[10, 10, 10]
+            weights = Int[5, 10, 15]
+            values  = Int[10, 10, 10]
             epsilon = 0.2
             opt_approx, chosen = ptas_knapsack(W, epsilon, weights, values)
             opt_exact, _ = exact_knapsack(W, weights, values)
@@ -198,8 +198,8 @@ end
 
         @testset "large epsilon still reasonable" begin
             W = 50
-            weights = Int64[10, 20, 30]
-            values  = Int64[60, 100, 120]
+            weights = Int[10, 20, 30]
+            values  = Int[60, 100, 120]
             epsilon = 0.5  # 50% approximation
             opt_approx, chosen = ptas_knapsack(W, epsilon, weights, values)
             opt_exact, _ = exact_knapsack(W, weights, values)
